Added norwegian WiFis from personal scanning+WiGLE
Not sure how common "common" is but I've found quite a few. Could be an idea for tools to support regexes for common guest SSIDs, then we'd be able to have just a few lines for variations like "guest", "gjest", ... . HSN-guest is being deprecated but you should still see it for a few years in probe-requests.